Conflicts:
	src/main/cz/fel/ds/database/model/Diet.java
	src/main/cz/fel/ds/database/model/Exercise.java
	src/main/cz/fel/ds/database/model/ExerciseToTrainingProgram.java
	src/main/cz/fel/ds/database/model/Food.java
	src/main/cz/fel/ds/database/model/Meal.java
	src/main/cz/fel/ds/database/model/MealSchedule.java
	src/main/cz/fel/ds/database/model/MealScheduleChange.java
	src/main/cz/fel/ds/database/model/MealToFood.java
	src/main/cz/fel/ds/database/model/MealType.java
	src/main/cz/fel/ds/database/model/MedicalRecord.java
	target/classes/cz/fel/ds/database/model/Diet.class
	target/classes/cz/fel/ds/database/model/Excercise.class

	target/classes/cz/fel/ds/database/model/ExcerciseToTrainingProgram.class
	target/classes/cz/fel/ds/database/model/Food.class
	target/classes/cz/fel/ds/database/model/Meal.class
	target/classes/cz/fel/ds/database/model/MealSchedule.class
	target/classes/cz/fel/ds/database/model/MealScheduleChange.class
	target/classes/cz/fel/ds/database/model/MealToFood.class
	target/classes/cz/fel/ds/database/model/MealType.class
	target/classes/cz/fel/ds/database/model/MedicalRecord.class
	target/test-classes/DAO/testDAO.class

+++ src/main/cz/fel/ds/database/model/MealToFood.java
@@ -1,7 +1,6 @@
 package cz.fel.ds.database.model;
 
-import javafx.beans.property.SimpleFloatProperty;
-
+import javafx.beans.property.SimpleDoubleProperty;
 import javax.persistence.*;
 import java.io.Serializable;
 
@@ -10,19 +9,19 @@ import java.io.Serializable;
  */
 @Entity
 @Table(name = "MealToFood")
-public class MealToFood implements Serializable
+public class MealToFood implements Serializable 
 {
-    private SimpleFloatProperty quantity = new SimpleFloatProperty();
+    private SimpleDoubleProperty quantity = new SimpleDoubleProperty();
     private Meal meal;
     private Food food;
 
     @Column(name = "quantity", length = 100)
-    public float getQuantity()
+    public double getQuantity() 
     {
         return quantity.get();
     }
 
-    public void setQuantity(float quantity)
+    public void setQuantity(double quantity)
     {
         this.quantity.set(quantity);
     }
